create generic Spectrum2DViewer
* mosviz's MosvizProfile2DView inherits from this now
