+++ jdaviz/configs/specviz/plugins/viewers.py
@@ -2,17 +2,25 @@ import warnings
 
 import numpy as np
 from astropy import table
+from functools import cached_property
 from matplotlib.colors import cnames
 from specutils import Spectrum1D
+from scipy.interpolate import interp1d
+from glue.core import BaseData
+from glue_jupyter.bqplot.image import BqplotImageView
 
-from jdaviz.core.events import SpectralMarksChangedMessage, LineIdentifyMessage
+from jdaviz.core.events import (AddDataToViewerMessage,
+                                RemoveDataFromViewerMessage,
+                                SpectralMarksChangedMessage,
+                                LineIdentifyMessage)
+from jdaviz.core.freezable_state import FreezableBqplotImageViewerState
 from jdaviz.core.registries import viewer_registry
 from jdaviz.core.marks import SpectralLine
 from jdaviz.core.linelists import load_preset_linelist, get_available_linelists
 from jdaviz.core.freezable_state import FreezableProfileViewerState
-from jdaviz.configs.default.plugins.viewers import JdavizProfileView
+from jdaviz.configs.default.plugins.viewers import JdavizViewerMixin, JdavizProfileView
 
-__all__ = ['Spectrum1DViewer']
+__all__ = ['Spectrum1DViewer', 'Spectrum2DViewer']
 
 
 @viewer_registry("spectrum-1d-viewer", label="1D Spectrum")
@@ -36,7 +44,6 @@ class Spectrum1DViewer(JdavizProfileView):
         super().__init__(*args, **kwargs)
         self.data_menu._obj.dataset.add_filter('is_spectrum')
         self.data_menu.layer.add_filter('not_trace')
-        self.data_menu._obj.dataset.add_filter('same_mosviz_row')
 
     @property
     def redshift(self):
@@ -248,3 +255,172 @@ class Spectrum1DViewer(JdavizProfileView):
     def set_plot_axes(self):
         super().set_plot_axes()
         self.figure.axes[1].num_ticks = 5
+
+
+@viewer_registry("spectrum-2d-viewer", label="2D Spectrum")
+class Spectrum2DViewer(JdavizViewerMixin, BqplotImageView):
+    # Due to limitations in CCDData and 2D data that has spectral and spatial
+    #  axes, the default conversion class must handle cubes
+    default_class = Spectrum1D
+
+    # categories: zoom resets, zoom, pan, subset, select tools, shortcuts
+    tools_nested = [
+                    ['mosviz:homezoom'],
+                    ['mosviz:boxzoom', 'mosviz:xrangezoom', 'jdaviz:yrangezoom'],
+                    ['mosviz:panzoom', 'mosviz:panzoom_x', 'jdaviz:panzoom_y'],
+                    ['bqplot:xrange'],
+                    ['jdaviz:sidebar_plot', 'jdaviz:sidebar_export']
+                ]
+
+    _state_cls = FreezableBqplotImageViewerState
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._subscribe_to_layers_update()
+        # Setup viewer option defaults
+        self.state.aspect = 'auto'
+
+        self.session.hub.subscribe(self, AddDataToViewerMessage,
+                                   handler=self._on_viewer_data_changed)
+        self.session.hub.subscribe(self, RemoveDataFromViewerMessage,
+                                   handler=self._on_viewer_data_changed)
+
+        for k in ('x_min', 'x_max'):
+            self.state.add_callback(k, self._handle_x_axis_orientation)
+
+        self.data_menu._obj.dataset.add_filter('is_2d_spectrum_or_trace')
+
+    @cached_property
+    def reference_spectral_axis(self):
+        return self.state.reference_data.get_object().spectral_axis.value
+
+    @cached_property
+    def pixel_to_world_interp(self):
+        pixels = range(len(self.reference_spectral_axis))
+        return interp1d(pixels, self.reference_spectral_axis)
+
+    def pixel_to_world_limits(self, limits):
+        if not len(limits) == 2:
+            raise ValueError("limits must be length 2")
+
+        pixels = np.arange(0, len(self.reference_spectral_axis))
+
+        # we'll use interpolation when possible, but also want to fit a line between
+        # the outermost edge of the data within the limits
+        line_edges_pix = np.array([max((min(pixels), min(limits))),
+                                   min((max(pixels), max(limits)))])
+        if line_edges_pix[0] > line_edges_pix[1]:
+            # then the limits are entirely out of range, so use the whole range
+            # when fitting the linear approximation
+            line_edges_pix = np.array([min(pixels), max(pixels)])
+        line_edges_world = self.pixel_to_world_interp(line_edges_pix)
+        line_coeffs = np.polyfit(line_edges_pix, line_edges_world, deg=1)
+
+        def pixel_to_world_line(pixel):
+            return line_coeffs[0] * pixel + line_coeffs[1]
+
+        def map_pixel_to_world(pixel):
+            if pixels[0] <= pixel <= pixels[-1]:
+                # interpolate directly
+                return float(self.pixel_to_world_interp(pixel))
+            else:
+                # use the line model to extrapolate
+                return pixel_to_world_line(pixel)
+
+        invert = (-1) ** sum((self.inverted_x_axis, limits[0] > limits[1]))
+        out_lims = list(map(map_pixel_to_world, limits))[::invert]
+
+        return out_lims
+
+    @cached_property
+    def world_to_pixel_interp(self):
+        pixels = range(len(self.reference_spectral_axis))
+        return interp1d(self.reference_spectral_axis, pixels)
+
+    def world_to_pixel_limits(self, limits):
+        if not len(limits) == 2:
+            raise ValueError("limits must be length 2")
+
+        # we'll use interpolation when possible, but also want to fit a line between
+        # the outermost edge of the data within the limits
+        line_edges_world = np.array([max((min(self.reference_spectral_axis), min(limits))),
+                                     min((max(self.reference_spectral_axis), max(limits)))])
+        if line_edges_world[0] > line_edges_world[1]:
+            # then the limits are entirely out of range, so use the whole range
+            # when fitting the linear approximation
+            line_edges_world = np.array([min(self.reference_spectral_axis),
+                                         max(self.reference_spectral_axis)])
+        line_edges_pixels = self.world_to_pixel_interp(line_edges_world)
+        line_coeffs = np.polyfit(line_edges_world, line_edges_pixels, deg=1)
+
+        def world_to_pixel_line(world):
+            return line_coeffs[0] * world + line_coeffs[1]
+
+        def map_world_to_pixel(world):
+            if min(self.reference_spectral_axis) <= world <= max(self.reference_spectral_axis):
+                # interpolate directly
+                return float(self.world_to_pixel_interp(world))
+            else:
+                # use the line model to extrapolate
+                return world_to_pixel_line(world)
+
+        invert = (-1) ** sum((self.inverted_x_axis, limits[0] > limits[1]))
+        out_lims = list(map(map_world_to_pixel, limits))[::invert]
+
+        return out_lims
+
+    def _on_viewer_data_changed(self, msg):
+        if msg.viewer_reference != self.reference:
+            return
+        # clear cached properties that are based on reference data - this is probably
+        # overly-conservative and we might be able to limit the clearing for only when
+        # reference data is changed (perhaps with a callback on the state for reference_data)
+        for attr in ('reference_spectral_axis', 'inverted_x_axis',
+                     'pixel_to_world_interp', 'world_to_pixel_interp'):
+            if attr in self.__dict__:
+                del self.__dict__[attr]
+        if len(self.data()):
+            self._handle_x_axis_orientation()
+
+    @cached_property
+    def inverted_x_axis(self):
+        return self.reference_spectral_axis[0] > self.reference_spectral_axis[-1]
+
+    def _handle_x_axis_orientation(self, *args):
+        x_scales = self.scales['x']
+        limits = [x_scales.min, x_scales.max]
+        limits_inverted = limits[0] > limits[1]
+        if limits_inverted == self.inverted_x_axis:
+            return
+        with x_scales.hold_sync():
+            x_scales.min = max(limits) if self.inverted_x_axis else min(limits)
+            x_scales.max = min(limits) if self.inverted_x_axis else max(limits)
+
+    def data(self, cls=None, statistic=None):
+        return [layer_state.layer.get_object(statistic=statistic,
+                                             cls=cls or self.default_class)
+                for layer_state in self.state.layers
+                if hasattr(layer_state, 'layer') and
+                isinstance(layer_state.layer, BaseData)]
+
+    def set_plot_axes(self):
+        self.figure.axes[0].label = "x: pixels"
+
+        self.figure.axes[1].label = "y: pixels"
+        self.figure.axes[1].tick_format = None
+        self.figure.axes[1].label_location = "middle"
+
+        # Sync with Spectrum1D viewer (that is also used by other viz).
+        # Make it so y axis label is not covering tick numbers.
+        self.figure.fig_margin["left"] = 95
+        self.figure.fig_margin["bottom"] = 60
+        self.figure.send_state('fig_margin')  # Force update
+        self.figure.axes[0].label_offset = "40"
+        self.figure.axes[1].label_offset = "-70"
+        # NOTE: with tick_style changed below, the default responsive ticks in bqplot result
+        # in overlapping tick labels.  For now we'll hardcode at 8, but this could be removed
+        # (default to None) if/when bqplot auto ticks react to styling options.
+        self.figure.axes[1].num_ticks = 8
+
+        for i in (0, 1):
+            self.figure.axes[i].tick_style = {'font-size': 15, 'font-weight': 600}
