+++ jdaviz/core/loaders/importers/importer.py
@@ -11,7 +11,7 @@ __all__ = ['BaseImporter', 'BaseImporterToDataCollection', 'BaseImporterToPlugin
 
 
 vid_map = {'spectrum-1d-viewer': '1D Spectrum',
-           'mosviz-profile-2d-viewer': '2D Spectrum'}
+           'spectrum-2d-viewer': '2D Spectrum'}
 
 
 class BaseImporter(PluginTemplateMixin):
