+++ jdaviz/core/marks.py
@@ -156,6 +156,7 @@ class PluginMark:
         if not self.auto_update_units:
             return
         if self.viewer.__class__.__name__ in ['Spectrum1DViewer',
+                                              'Spectrum2DViewer',
                                               'CubevizProfileView',
                                               'MosvizProfileView',
                                               'MosvizProfile2DView']:
